+++ rotkehlchen/chain/ethereum/modules/yearn/constants.py
@@ -4,8 +4,10 @@ BLOCKS_PER_YEAR = 2425846
 
 CPT_YEARN_V1: Final = 'yearn-v1'
 CPT_YEARN_V2: Final = 'yearn-v2'
+CPT_YEARN_V3: Final = 'yearn-v3'
 CPT_YGOV: Final = 'ygov'
 
 YEARN_LABEL_V1: Final = 'Yearn V1'
 YEARN_LABEL_V2: Final = 'Yearn V2'
+YEARN_LABEL_V3: Final = 'Yearn V3'
 YEARN_ICON: Final = 'yearn_vaults.svg'